+++ src/components/App/index.js
@@ -1,0 +1,1 @@
+export { default } from "./App.jsx"
