+++ src/components/LoginForm/style.css
@@ -1,83 +1,94 @@
-.login-form{
-	width: 502px;
-	height: 473px;
-	border-radius: 3px;
-	background-color: #ffffff;
-	display: -webkit-flex;
-	display: -moz-flex;
-	display: -ms-flex;
-	display: -o-flex;
-	display: flex;
-	-webkit-flex-direction: column;
-	-moz-flex-direction: column;
-	-ms-flex-direction: column;
-	-o-flex-direction: column;
-	flex-direction: column;
-	justify-content: space-between;
-	padding: 52px;
+.login-form {
+    width: 502px;
+    height: 473px;
+    border-radius: 3px;
+    background-color: #ffffff;
+    display: -webkit-flex;
+    display: -moz-flex;
+    display: -ms-flex;
+    display: -o-flex;
+    display: flex;
+    -webkit-flex-direction: column;
+    -moz-flex-direction: column;
+    -ms-flex-direction: column;
+    -o-flex-direction: column;
+    flex-direction: column;
+    justify-content: space-between;
+    padding: 52px;
 }
-.login-form h2{
-	margin: 0;
-	color: #323232;
-	font-size: 36px;
-	font-weight: 700;
+
+.login-form h2 {
+    margin: 0;
+    color: #323232;
+    font-size: 36px;
+    font-weight: 700;
 }
-.login-form__button{
-	width: 161px;
-	height: 51px;
-	border-radius: 4px;
-	border:none;
-	background-color: #ffc617;
-	align-self: flex-end;
-	font-size: 21px;
-	font-weight: 400;
-	outline:0;
-	cursor: pointer;
-	transition: .2s
+
+.login-form__button {
+    width: 161px;
+    height: 51px;
+    border-radius: 4px;
+    border: none;
+    background-color: #ffc617;
+    align-self: flex-end;
+    font-size: 21px;
+    font-weight: 400;
+    outline: 0;
+    cursor: pointer;
+    transition: .2s
 }
-.login-form__button:hover{
-	background-color: #B28A10;
+
+.login-form__button:hover {
+    background-color: #B28A10;
 }
-.login-form__input-box{
-	display: -webkit-flex;
-	display: -moz-flex;
-	display: -ms-flex;
-	display: -o-flex;
-	display: flex;
-	-webkit-flex-direction: column;
-	-moz-flex-direction: column;
-	-ms-flex-direction: column;
-	-o-flex-direction: column;
-	flex-direction: column;
+
+.login-form__input-box {
+    display: -webkit-flex;
+    display: -moz-flex;
+    display: -ms-flex;
+    display: -o-flex;
+    display: flex;
+    -webkit-flex-direction: column;
+    -moz-flex-direction: column;
+    -ms-flex-direction: column;
+    -o-flex-direction: column;
+    flex-direction: column;
 }
-.login-form__input-box span{
-	color: #747474;
-	font-size: 15px;
-	font-weight: 400;
+
+.login-form__input-box span {
+    color: #747474;
+    font-size: 15px;
+    font-weight: 400;
 }
-.login-form__input-box input{
-	margin: 1px;
-	height: 34px;
-	color: #323232;
-	font-size: 16px;
-	font-weight: 400;
-	border:none;
-	outline:0;
-	border-bottom: 2px solid #eaeaea;
-	transition: .15s;
+
+.login-form__input-box input {
+    margin: 1px;
+    height: 34px;
+    color: #323232;
+    font-size: 16px;
+    font-weight: 400;
+    border: none;
+    outline: 0;
+    border-bottom: 2px solid #eaeaea;
+    transition: .15s;
 }
-.login-form__input-box input:focus{
-	border-bottom: 2px solid #1473e6;
+
+.login-form__input-box input:focus {
+    border-bottom: 2px solid #1473e6;
 }
-.login-form__new{
-	font-size: 16px;
-	font-weight: 400;
-	color: #505050;
+
+.login-form__new {
+    font-size: 16px;
+    font-weight: 400;
+    color: #505050;
 }
-.login-form__new_blue{
-	color: #1473e6;
-	cursor: pointer;
+
+.login-form__new_blue {
+    color: #1473e6;
+    cursor: pointer;
+    text-decoration: none;
 }
-.login-form__new_blue:hover{
-	text-decoration: underline;
+
+.login-form__new_blue:hover {
+    text-decoration: underline;
 }
