+++ src/components/NavList/style.css
@@ -7,14 +7,16 @@ nav ul {
     display: flex;
 }
 
-nav ul li {
+nav ul li a {
     margin-right: 5px;
     padding: 7px;
     border-radius: 5px;
     cursor: pointer;
     transition: .3s;
+    text-decoration: none;
+    color: #212121;
 }
 
-nav ul li:hover {
+nav ul li a:hover {
     background-color: #ECEEED;
 }
