+++ src/components/RegForm/index.js
@@ -1,0 +1,1 @@
+export { default } from "./RegForm.jsx"
